Add support for UDH for MT (#162)
Add optional support for splitting messages using UDH instead of the
message payload TLV. The SMPP spec says message payload should be used
for long messages, but not all SMSCs support this. For those connections
we need to use UDH for splitting. We will then receive one reference
code for each part, as well as one delivery receipt for each part.
SmppResponse now has a list of references for this reason.