Fix long messages with UDH and unicode
Messages that need to be split, being sent to an SMSC that needs UDH
splitting of messagess, were previously being split the wrong way.

We only allocated an odd number of bytes for each part, which does not
fit well with UCS, which has two bytes per character. That is now fixed,
so there's an even number of characters in each part.